+++ NanoRpcSharp/Messages/MraiToRawRequest.cs
@@ -1,16 +1,18 @@
 ﻿namespace NanoRpcSharp.Messages
 {
+    using System.Numerics;
+
     /// <summary>
     /// Multiply an Mrai amount by the Mrai ratio.
     /// </summary>
     public class MraiToRawRequest : RequestBase<BigIntegerAmount>
     {
-        public MraiToRawRequest(decimal amount)
+        public MraiToRawRequest(BigInteger amount)
             : base("mrai_to_raw")
         {
             this.Amount = amount;
         }
 
-        public decimal Amount { get; set; }
+        public BigInteger Amount { get; set; }
     }
 }
