+++ NanoRpcSharp/Messages/RaiToRawRequest.cs
@@ -1,16 +1,18 @@
 ﻿namespace NanoRpcSharp.Messages
 {
+    using System.Numerics;
+
     /// <summary>
     /// Multiply an rai amount by the rai ratio.
     /// </summary>
     public class RaiToRawRequest : RequestBase<BigIntegerAmount>
     {
-        public RaiToRawRequest(decimal amount)
+        public RaiToRawRequest(BigInteger amount)
             : base("rai_to_raw")
         {
             this.Amount = amount;
         }
 
-        public decimal Amount { get; set; }
+        public BigInteger Amount { get; set; }
     }
 }
